Use single mutex to protect node data and children
Each node in ruletree have two RWMutex structures that is 48 bytes per node.
This change uses a single Mutex to protect both children and data
structures which uses only 8 bytes of memory. This is a trade-off
between concurrency and memory usage, but in practice the contention
is low and the memory savings are significant.